requirements_flux.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
alembic==1.18.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
av==16.1.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
color-matcher==0.6.0
colour-science==0.4.6
comfy-aimdo==0.1.7
comfy-kitchen==0.2.7
comfyui-embedded-docs==0.4.0
comfyui-manager==4.1b1
comfyui-workflow-templates-core==0.3.124
comfyui-workflow-templates-media-api==0.3.47
comfyui-workflow-templates-media-image==0.3.77
comfyui-workflow-templates-media-other==0.3.106
comfyui-workflow-templates-media-video==0.3.43
comfyui_frontend_package==1.37.11
comfyui_workflow_templates==0.8.31
contourpy==1.3.2
cryptography==46.0.4
cuda-bindings==13.0.3
cuda-pathfinder==1.2.2
cycler==0.12.1
ddt==1.7.2
dill==0.4.1
distro==1.9.0
docutils==0.22.4
easydict==1.13
einops==0.8.2
exceptiongroup==1.3.1
filelock==3.20.3
filetype==1.2.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
gdown==5.2.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
google-genai==1.69.0
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
iopath==0.1.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
langchain-core==1.2.23
langchain-google-genai==4.2.1
langsmith==0.7.22
lazy_loader==0.4
llvmlite==0.46.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
mss==10.1.0
multidict==6.7.1
networkx==3.4.2
numba==0.63.1
numpy==2.2.6
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.15.1.9
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
omegaconf==2.3.0
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.90
orjson==3.11.7
packaging==26.0
pandas==2.3.3
piexif==1.1.3
pillow==12.1.0
pixeloe==0.1.4
platformdirs==4.5.1
pooch==1.9.0
portalocker==3.2.0
propcache==0.4.1
psutil==7.2.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.11.0
PyMatting==1.1.15
PyNaCl==1.6.2
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
rembg==2.0.69
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
rpds-py==0.30.0
safetensors==0.7.0
SAM-2 @ git+https://github.com/facebookresearch/sam2@2b90b9f5ceec907a1c18123530e92e794ad901a4
scikit-image==0.25.2
scipy==1.15.3
segment-anything==1.0
sentencepiece==0.2.1
shellingham==1.5.4
simsimd==6.5.12
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
spandrel==0.4.1
SQLAlchemy==2.0.46
stringzilla==4.6.0
sympy==1.14.0
tenacity==9.1.4
tifffile==2025.5.10
timm==1.0.24
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.0
torch==2.10.0+cu130
torchaudio==2.10.0+cu130
torchsde==0.2.6
torchvision==0.25.0+cu130
tqdm==4.67.1
trampoline==0.1.2
transformers==4.57.6
transparent-background==1.3.4
triton==3.6.0
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uv==0.9.29
websockets==16.0
wget==3.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
